api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "lightrag-hku[api]>=1.3.7",
    "nano-vectordb>=0.0.4.3",
    "pydantic-ai>=0.2.6",
]
